+++ demo/lib/screen/home_screen.dart
@@ -662,6 +662,41 @@ class PlutoContributors extends StatelessWidget {
               launchUrl('https://github.com/Milad-Akarie');
             },
           ),
+          PlutoContributorTile(
+            name: 'Verry',
+            linkTitle: 'Github',
+            onTapLink: () {
+              launchUrl('https://github.com/novas1r1');
+            },
+          ),
+          PlutoContributorTile(
+            name: 'coda538',
+            linkTitle: 'Github',
+            onTapLink: () {
+              launchUrl('https://github.com/coda538');
+            },
+          ),
+          PlutoContributorTile(
+            name: 'billyio',
+            linkTitle: 'Github',
+            onTapLink: () {
+              launchUrl('https://github.com/billyio');
+            },
+          ),
+          PlutoContributorTile(
+            name: 'Mehmet',
+            linkTitle: 'Github',
+            onTapLink: () {
+              launchUrl('https://github.com/mehmetkalayci');
+            },
+          ),
+          PlutoContributorTile(
+            name: 'Tautvydas Šidlauskas',
+            linkTitle: 'Github',
+            onTapLink: () {
+              launchUrl('https://github.com/sidlatau');
+            },
+          ),
           PlutoContributorTile.invisible(
             name: 'And you.',
             linkTitle: 'Github',