+++ test/src/pluto_grid_configuration_test.dart
@@ -644,5 +644,23 @@ void main() {
 
       expect(locale.loadingText, 'Laster');
     });
+
+    test('german 이 호출 되어야 한다.', () {
+      const locale = PlutoGridLocaleText.german();
+
+      expect(locale.loadingText, 'Lädt');
+    });
+
+    test('turkish 이 호출 되어야 한다.', () {
+      const locale = PlutoGridLocaleText.turkish();
+
+      expect(locale.loadingText, 'Yükleniyor');
+    });
+
+    test('japanese 이 호출 되어야 한다.', () {
+      const locale = PlutoGridLocaleText.japanese();
+
+      expect(locale.loadingText, 'にゃ〜');
+    });
   });
 }