+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## [7.0.2] - 2023. 5. 13
+
+* Add Japanese, Turkish languages.
+* Fix Intl version.
+
+## [7.0.1] - 2023. 5. 11
+
+* Fix SDK version.
+
+## [7.0.0] - 2023. 5. 11
+
+* Compatible with Flutter 3.10.
+
+## [6.0.4] - 2023. 3. 7
+
+* Add Scrollbar properties. https://github.com/bosskmk/pluto_grid/issues/743
+
+## [6.0.3] - 2023. 3. 6
+
+* Add locales.(German, French)
+* Fix row group expanded bug. https://github.com/bosskmk/pluto_grid/issues/775
+
+## [6.0.2] - 2023. 2. 2
+
+* Fix filtering bug of number.
+
+## [6.0.1] - 2023. 1. 26
+
+* Fix rendering issue when tapping another cell after editing a cell.
+
+## [6.0.0] - 2023. 1. 26
+
+* Updated for flutter 3.7 version.
+
 ## [5.4.9] - 2022. 12. 15
 
 * Fix sortIdx bug.